Migrate SL to use DynamicObjectLibrary and introduce SLObject subclass.
